add test for listFiles
